ref(nextjs): Move pages router instrumentation into separate folder
