+++ packages/nextjs/src/common/index.ts
@@ -1,14 +1,14 @@
-export { wrapGetStaticPropsWithSentry } from './wrapGetStaticPropsWithSentry';
-export { wrapGetInitialPropsWithSentry } from './wrapGetInitialPropsWithSentry';
-export { wrapAppGetInitialPropsWithSentry } from './wrapAppGetInitialPropsWithSentry';
-export { wrapDocumentGetInitialPropsWithSentry } from './wrapDocumentGetInitialPropsWithSentry';
-export { wrapErrorGetInitialPropsWithSentry } from './wrapErrorGetInitialPropsWithSentry';
-export { wrapGetServerSidePropsWithSentry } from './wrapGetServerSidePropsWithSentry';
+export { wrapGetStaticPropsWithSentry } from './pages-router-instrumentation/wrapGetStaticPropsWithSentry';
+export { wrapGetInitialPropsWithSentry } from './pages-router-instrumentation/wrapGetInitialPropsWithSentry';
+export { wrapAppGetInitialPropsWithSentry } from './pages-router-instrumentation/wrapAppGetInitialPropsWithSentry';
+export { wrapDocumentGetInitialPropsWithSentry } from './pages-router-instrumentation/wrapDocumentGetInitialPropsWithSentry';
+export { wrapErrorGetInitialPropsWithSentry } from './pages-router-instrumentation/wrapErrorGetInitialPropsWithSentry';
+export { wrapGetServerSidePropsWithSentry } from './pages-router-instrumentation/wrapGetServerSidePropsWithSentry';
 export { wrapServerComponentWithSentry } from './wrapServerComponentWithSentry';
 export { wrapRouteHandlerWithSentry } from './wrapRouteHandlerWithSentry';
-export { wrapApiHandlerWithSentryVercelCrons } from './wrapApiHandlerWithSentryVercelCrons';
+export { wrapApiHandlerWithSentryVercelCrons } from './pages-router-instrumentation/wrapApiHandlerWithSentryVercelCrons';
 export { wrapMiddlewareWithSentry } from './wrapMiddlewareWithSentry';
-export { wrapPageComponentWithSentry } from './wrapPageComponentWithSentry';
+export { wrapPageComponentWithSentry } from './pages-router-instrumentation/wrapPageComponentWithSentry';
 export { wrapGenerationFunctionWithSentry } from './wrapGenerationFunctionWithSentry';
 export { withServerActionInstrumentation } from './withServerActionInstrumentation';
 // eslint-disable-next-line deprecation/deprecation
