+++ packages/nextjs/src/common/pages-router-instrumentation/wrapApiHandlerWithSentryVercelCrons.ts
@@ -1,7 +1,7 @@
 import { captureCheckIn } from '@sentry/core';
 import type { NextApiRequest } from 'next';
 
-import type { VercelCronsConfig } from './types';
+import type { VercelCronsConfig } from '../types';
 
 type EdgeRequest = {
   nextUrl: URL;
