+++ packages/nextjs/src/common/pages-router-instrumentation/wrapAppGetInitialPropsWithSentry.ts
@@ -1,7 +1,7 @@
 import type App from 'next/app';
 
-import { isBuild } from './utils/isBuild';
-import { withErrorInstrumentation, withTracedServerSideDataFetcher } from './utils/wrapperUtils';
+import { isBuild } from '../utils/isBuild';
+import { withErrorInstrumentation, withTracedServerSideDataFetcher } from '../utils/wrapperUtils';
 
 type AppGetInitialProps = (typeof App)['getInitialProps'];
