+++ packages/nextjs/src/common/pages-router-instrumentation/wrapErrorGetInitialPropsWithSentry.ts
@@ -1,20 +1,24 @@
-import type { NextPage } from 'next';
+import type { NextPageContext } from 'next';
+import type { ErrorProps } from 'next/error';
 
-import { isBuild } from './utils/isBuild';
-import { withErrorInstrumentation, withTracedServerSideDataFetcher } from './utils/wrapperUtils';
+import { isBuild } from '../utils/isBuild';
+import { withErrorInstrumentation, withTracedServerSideDataFetcher } from '../utils/wrapperUtils';
 
-type GetInitialProps = Required<NextPage>['getInitialProps'];
+type ErrorGetInitialProps = (context: NextPageContext) => Promise<ErrorProps>;
 
 /**
- * Create a wrapped version of the user's exported `getInitialProps` function
+ * Create a wrapped version of the user's exported `getInitialProps` function in
+ * a custom error page ("_error.js").
  *
- * @param origGetInitialProps The user's `getInitialProps` function
+ * @param origErrorGetInitialProps The user's `getInitialProps` function
  * @param parameterizedRoute The page's parameterized route
  * @returns A wrapped version of the function
  */
-export function wrapGetInitialPropsWithSentry(origGetInitialProps: GetInitialProps): GetInitialProps {
-  return new Proxy(origGetInitialProps, {
-    apply: async (wrappingTarget, thisArg, args: Parameters<GetInitialProps>) => {
+export function wrapErrorGetInitialPropsWithSentry(
+  origErrorGetInitialProps: ErrorGetInitialProps,
+): ErrorGetInitialProps {
+  return new Proxy(origErrorGetInitialProps, {
+    apply: async (wrappingTarget, thisArg, args: Parameters<ErrorGetInitialProps>) => {
       if (isBuild()) {
         return wrappingTarget.apply(thisArg, args);
       }
@@ -29,35 +33,35 @@ export function wrapGetInitialPropsWithSentry(origGetInitialProps: GetInitialPro
       // span with each other when there are no req or res objects, we simply do not trace them at all here.
       if (req && res) {
         const tracedGetInitialProps = withTracedServerSideDataFetcher(errorWrappedGetInitialProps, req, res, {
-          dataFetcherRouteName: context.pathname,
+          dataFetcherRouteName: '/_error',
           requestedRouteName: context.pathname,
           dataFetchingMethodName: 'getInitialProps',
         });
 
         const {
-          data: initialProps,
+          data: errorGetInitialProps,
           baggage,
           sentryTrace,
         }: {
-          data: {
+          data: ErrorProps & {
             _sentryTraceData?: string;
             _sentryBaggage?: string;
           };
           baggage?: string;
           sentryTrace?: string;
-        } = (await tracedGetInitialProps.apply(thisArg, args)) ?? {}; // Next.js allows undefined to be returned from a getInitialPropsFunction.
+        } = await tracedGetInitialProps.apply(thisArg, args);
 
         // The Next.js serializer throws on undefined values so we need to guard for it (#12102)
         if (sentryTrace) {
-          initialProps._sentryTraceData = sentryTrace;
+          errorGetInitialProps._sentryTraceData = sentryTrace;
         }
 
         // The Next.js serializer throws on undefined values so we need to guard for it (#12102)
         if (baggage) {
-          initialProps._sentryBaggage = baggage;
+          errorGetInitialProps._sentryBaggage = baggage;
         }
 
-        return initialProps;
+        return errorGetInitialProps;
       } else {
         return errorWrappedGetInitialProps.apply(thisArg, args);
       }
