+++ packages/nextjs/src/common/pages-router-instrumentation/wrapGetStaticPropsWithSentry.ts
@@ -1,7 +1,7 @@
 import type { GetStaticProps } from 'next';
 
-import { isBuild } from './utils/isBuild';
-import { callDataFetcherTraced, withErrorInstrumentation } from './utils/wrapperUtils';
+import { isBuild } from '../utils/isBuild';
+import { callDataFetcherTraced, withErrorInstrumentation } from '../utils/wrapperUtils';
 
 type Props = { [key: string]: unknown };
